Refactor score_candidates -> dynamic_oracle
